Upgrade Undertow version to 2.2.19 (CVE-2022-2053)
When a POST request comes through AJP and the request exceeds the
max-post-size limit (maxEntitySize), Undertow's AjpServerRequestConduit
implementation closes a connection without sending any response to the
client/proxy. This behavior results in that a front-end proxy marking
the backend worker (application server) as an error state and not forward
requests to the worker for a while. CWE-400 / CVE-2022-2053.

Signed-off-by: Thibault Meyer <meyer.thibault@gmail.com>